+++ src/helpers/encode-link.ts
@@ -16,18 +16,13 @@ const encodeLink: TEncodeLink = (
 
   const linkKey = encodeBase58(link.linkKey)
   const transferId = encodeBase58(link.transferId) // string -> hex -> base58 for shorter string
-  const symbol = defineTokenSymbol(
-    link.tokenType as TTokenType,
-    link.chainId,
-    tokenAddress
-  )
 
   if (link.senderSig) {
     const sig = encodeBase58(link.senderSig)
-    return `${claimHost}/#/${symbol}?k=${linkKey}&sg=${sig}&i=${transferId}&c=${link.chainId}&v=2`
+    return `${claimHost}/#/code?k=${linkKey}&sg=${sig}&i=${transferId}&c=${link.chainId}&v=3`
   } else if (link.sender) {
     const sender = encodeBase58(link.sender)
-    return `${claimHost}/#/${symbol}?k=${linkKey}&s=${sender}&c=${link.chainId}&v=2`
+    return `${claimHost}/#/code?k=${linkKey}&s=${sender}&c=${link.chainId}&v=3`
   }
 }