+++ src/helpers/index.ts
@@ -1,27 +1,25 @@
 import defineNetworkName from './define-network-name'
-import defineJSONRpcUrl from './define-json-rpc-url'
 import encodeLink from './encode-link'
 import decodeLink from './decode-link'
 import defineApiHost from './define-api-host'
 import parseLink from './parse-link'
 import defineEscrowAddress from './define-escrow-address'
-import defineEscrowAddressByTokenSymbol from './define-escrow-address-by-token-symbol'
 import defineTokenType from './define-token-type'
 import defineTokenSymbol from './define-token-symbol'
 import defineIfEscrowAddressIsCorrect from './define-if-escrow-address-is-correct'
 import createQueryString from './create-query-string'
 import updateOperations from './update-operations'
 import parseQueryParams from './parse-query-params'
+import defineDomain from './define-domain'
 
 export {
-  defineJSONRpcUrl,
+  defineDomain,
   updateOperations,
   parseQueryParams,
   createQueryString,
   defineIfEscrowAddressIsCorrect,
   defineTokenSymbol,
   defineNetworkName,
-  defineEscrowAddressByTokenSymbol,
   defineApiHost,
   defineTokenType,
   encodeLink,