+++ src/utils/get-nonce/index.ts
@@ -4,11 +4,12 @@ const getNonce = (
   from: string,
   transferId: string,
   amount: string,
-  expiration: string
+  expiration: string,
+  feeAmount: string
 ) => {
   return ethers.solidityPackedKeccak256(
-    ['address', 'address', 'uint256', 'uint128'],
-    [from, transferId, amount, expiration]
+    ['address', 'address', 'uint256', 'uint120', 'uint128'],
+    [from, transferId, amount, expiration, feeAmount]
   )
 }